fix : 출력 결과가 소수자리까지 나오도록 수정
소수자리 표현을 위해 곱해주는 상수를 int로 두어서 정수가 출력되는 버그 수정

+++ src/main/java/coordinate/view/ResultView.java
@@ -1,9 +1,11 @@
 package coordinate.view;
 
+import java.text.DecimalFormat;
+
 public class ResultView {
 
     private static final String DISTANCE_SCRIPT = "두 점 사이 거리는 ";
-    public static final int DECIMAL_PLACE = 1_000_000;
+    public static final double DECIMAL_PLACE = 1_000_000;
 
     public static void printResult(String graph) {
         System.out.println(graph);
